pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Self-evolving multi-agent academic research system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.30.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    "pydantic>=2.0",
    "structlog>=24.0",
    "httpx>=0.27.0",
    "pypdf>=4.0",
    "python-dotenv>=1.0",
    "fastembed>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
